refactor(elasticsearch): consolidate and reorganize Elasticsearch client implementation
This commit removes the existing Elasticsearch test and module files, consolidating the client implementations for versions v7, v8, and v9 into a new internal package. The new structure enhances modularity and type safety, while also introducing a unified interface for Elasticsearch operations. Additionally, comprehensive unit tests for CRUD operations and client functionality are added to ensure robust integration within the storage layer.
