--- internal/storage/elasticsearch/go.mod
@@ -1,23 +1,0 @@
-module trpc.group/trpc-go/trpc-agent-go/internal/storage/elasticsearch
-
-go 1.24.1
-
-require (
-	github.com/elastic/go-elasticsearch/v7 v7.17.10
-	github.com/elastic/go-elasticsearch/v8 v8.19.0
-	github.com/elastic/go-elasticsearch/v9 v9.1.0
-	github.com/stretchr/testify v1.10.0
-)
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/elastic/elastic-transport-go/v8 v8.7.0 // indirect
-	github.com/go-logr/logr v1.4.2 // indirect
-	github.com/go-logr/stdr v1.2.2 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	go.opentelemetry.io/auto/sdk v1.1.0 // indirect
-	go.opentelemetry.io/otel v1.35.0 // indirect
-	go.opentelemetry.io/otel/metric v1.35.0 // indirect
-	go.opentelemetry.io/otel/trace v1.35.0 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
